added recommendation tests
