+++ .gitignore
@@ -23,3 +23,5 @@ __pycache__/
 *.pyc
 
 env
+env2
+env3
