+++ npkreadings/models.py
@@ -1,5 +1,5 @@
 from django.db import models
-from recommendations.models import Recommendations
+from recommendation.models import Recommendation
 
 
 # Create your models here.
@@ -8,7 +8,7 @@ class NpkReading(models.Model):
     farmer_id = models.IntegerField()        
     sensor_id = models.IntegerField()
     npk_reading = models.IntegerField()     
-    recommendation_id = models.ForeignKey(Recommendations, on_delete=models.CASCADE,related_name='NpkReading' )
+    recommendation_id = models.ForeignKey(Recommendation, on_delete=models.CASCADE,related_name='NpkReading' )
     
     def __str__(self):
         return f"NpkReading {self.id}: Farmer {self.farmer_id}, Sensor {self.sensor_id}, Reading {self.npk_reading}"
