+++ recommendation/apps.py
@@ -4,3 +4,6 @@ from django.apps import AppConfig
 class RecommendationsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'recommendation'
+    
+    def ready(self):
+        import recommendation.signals
