+++ requirements.txt
@@ -1,4 +1,3 @@
-django ~= 4.0
 python-dotenv ~= 0.19
 requests ~= 2.27
 djangorestframework==3.15.2
@@ -23,3 +22,4 @@ requests-oauthlib==2.0.0
 black==24.8.0
 dj-database-url==2.2.0
 pytz==2024.2
+Django==5.0
